dom/serviceworkers/test/test_controller.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1002570 - test controller instance.</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script src="utils.js"></script>
<script class="testbody" type="text/javascript">

  var content;
  var iframe;
  var registration;

  function simpleRegister() {
    // We use the control scope for the less specific registration. The window will register a worker on controller/
    return navigator.serviceWorker.register("worker.js", { scope: "./control" })
      .then(swr => waitForState(swr.installing, 'activated', swr))
      .then(swr => registration = swr);
  }

  function unregister() {
    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    }, function(e) {
      dump("Unregistering the SW failed: " + e + "\n");
    });
  }

  function testController() {
    var p = new Promise(function(resolve, reject) {
      window.onmessage = function(e) {
        if (e.data.status == "ok") {
          ok(e.data.result, e.data.message);
        } else if (e.data.status == "done") {
          window.onmessage = null;
          content.removeChild(iframe);
          resolve();
        }
      }
    });

    content = document.getElementById("content");
    ok(content, "Parent exists.");

    iframe = document.createElement("iframe");
    iframe.setAttribute('src', "controller/index.html");
    content.appendChild(iframe);

    return p;
  }

  // This document just flips the prefs and opens the iframe for the actual test.
  function runTest() {
    simpleRegister()
      .then(testController)
      .then(unregister)
      .then(function() {
        SimpleTest.finish();
      }).catch(function(e) {
        ok(false, "Some test failed with error " + e);
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
